dom/indexedDB/test/test_file_resurrection_delete.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const READ_WRITE = IDBTransaction.READ_WRITE;

    const name = window.location.pathname;
    const description = "My Test Database";

    const objectStoreName = "Blobs";

    const fileData = { key: 1, file: getRandomFile("random.bin", 100000) };

    {
      let request = indexedDB.open(name, 1, description);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });

      objectStore.add(fileData.file, fileData.key);

      event = yield;

      is(event.type, "success", "Got correct event type");

      let trans = db.transaction([objectStoreName], READ_WRITE);
      objectStore = trans.objectStore(objectStoreName);

      objectStore.delete(fileData.key);

      trans.oncomplete = grabEventAndContinueHandler;
      event = yield;

      is(getFileDBRefCount(name, 1), 0, "Correct db ref count");

      trans = db.transaction([objectStoreName], READ_WRITE);
      objectStore = trans.objectStore(objectStoreName);

      request = objectStore.add(fileData.file, fileData.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      trans.oncomplete = grabEventAndContinueHandler;
      event = yield;

      is(getFileDBRefCount(name, 1), 1, "Correct db ref count");

      fileData.file = null;
    }

    scheduleGC();
    yield;

    is(getFileRefCount(name, 1), 0, "Correct ref count");

    {
      let request = indexedDB.open(name, 1, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      is(event.type, "success", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let trans = db.transaction([objectStoreName], READ_WRITE);
      objectStore = trans.objectStore(objectStoreName);

      request = objectStore.get(fileData.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let result = event.target.result;
      ok(result, "Got result");

      objectStore.delete(fileData.key);

      trans.oncomplete = grabEventAndContinueHandler;
      event = yield;

      is(getFileDBRefCount(name, 1), 0, "Correct db ref count");

      trans = db.transaction([objectStoreName], READ_WRITE);
      objectStore = trans.objectStore(objectStoreName);

      request = objectStore.add(result, fileData.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      trans.oncomplete = grabEventAndContinueHandler;
      event = yield;

      is(getFileDBRefCount(name, 1), 1, "Correct db ref count");

      event = null;
      result = null;
    }

    scheduleGC();
    yield;

    is(getFileRefCount(name, 1), 0, "Correct ref count");

    finishTest();
    yield;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
